requirements.txt
fastapi
uvicorn
openai
faiss-cpu
sentence-transformers
python-dotenv
azure-storage-blob